create usersMiddleware
