+++ teste2e/cypress/support/index.js
@@ -16,7 +16,6 @@
 // Import @testing-library/cypress commands:
 import "@testing-library/cypress/add-commands";
 // Import commands.js using ES2015 syntax:
-import "./general/commands";
 import "./commands";
 import "./core/commands";
 import "./ticketvote/commands";
