+++ teste2e/cypress/support/pi/commands.js
@@ -2,7 +2,6 @@ import "@testing-library/cypress/add-commands";
 import { createMiddleware } from "../core/commands";
 import { repliers, API_BASE_URL } from "./api";
 import { makeProposal } from "../../utils";
-import faker from "faker";
 
 Cypress.Commands.add(
   "piMiddleware",
@@ -30,19 +29,7 @@ Cypress.Commands.add("useProposalDetailSuite", ({ token, status, user }) => {
     amountByStatus: { [status]: 1 }
   });
   cy.piMiddleware("summaries", { amountByStatus: { [status]: 1 } });
-  cy.middleware("users.users", {
-    body: {
-      totalmatches: 1,
-      totalusers: 10,
-      users: [
-        {
-          id: faker.random.uuid(),
-          username: faker.internet.userName(),
-          email: faker.internet.email()
-        }
-      ]
-    }
-  });
+  cy.usersMiddleware(null, { amount: 1 }, {}, ["publickey"]);
   cy.recordsMiddleware("setstatus", {
     user
   });
