+++ teste2e/cypress/support/users/api.js
@@ -7,8 +7,10 @@ import {
   Identity,
   USER_TYPE_UNPAID
 } from "./generate";
+import faker from "faker";
 
 export const API_BASE_URL = "/api/v1/user";
+export const API_USERS_BASE_URL = "/api/v1/users";
 
 /**
  * loginReply is the replier to the Login command.
@@ -95,4 +97,35 @@ export const repliers = {
   "payments/registration": paymentsRegistrationReply,
   "payments/paywall": paymentsPaywallReply,
   "payments/credits": paymentsCreditsReply
+};
+
+/**
+ * usersReply represents the data of /api/v1/users endpoint
+ * It currently returns empty data since it is serving the data for downloading
+ * and we just check the existence of the downloaded file.
+ *
+ * @param {Object} { requestParams, testParams }
+ * @returns {Object}
+ */
+function usersReply({
+  requestParams: { publickey },
+  testParams: { amount = 0 }
+}) {
+  const users = [];
+  for (let i = 0; i < amount; i++) {
+    users.push({
+      id: faker.datatype.uuid(),
+      username: faker.internet.userName(),
+      email: faker.internet.email()
+    });
+  }
+  return {
+    totalmatches: amount,
+    totalusers: amount + 10,
+    users
+  };
+}
+
+export const usersRepliers = {
+  index: usersReply
 };
