+++ teste2e/cypress/support/users/commands.js
@@ -1,17 +1,21 @@
 import "@testing-library/cypress/add-commands";
 import "../www/commands";
-import { API_BASE_URL, repliers } from "./api";
+import { API_BASE_URL, API_USERS_BASE_URL, repliers, usersRepliers } from "./api";
 import { createMiddleware } from "../core/commands";
 import {
   USER_TYPE_UNPAID,
   USER_TYPE_NO_CREDITS,
-  USER_TYPE_NO_LOGIN,
-  User
+  USER_TYPE_NO_LOGIN
 } from "./generate";
 
 Cypress.Commands.add(
   "userMiddleware",
   createMiddleware({ packageName: "user", repliers, baseUrl: API_BASE_URL })
+);
+
+Cypress.Commands.add(
+  "usersMiddleware",
+  createMiddleware({ packageName: "users", repliers: usersRepliers, baseUrl: API_USERS_BASE_URL })
 );
 
 Cypress.Commands.add(
